Decouple ProcessData from the main project to avoid resources-processes to link to libadwaita
