+++ lib/process_data/Cargo.toml
@@ -1,0 +1,24 @@
+[package]
+name = "process-data"
+version = "1.0.3"
+authors = ["nokyan <nokyan@tuta.io>"]
+edition = "2021"
+
+[profile.dev]
+opt-level = 1
+
+[profile.release]
+opt-level = 3
+strip = true
+lto = true
+
+[dependencies]
+unescape = "0.1.0"
+rmp-serde = "1.1.2"
+anyhow = "1.0.71"
+async-std = { version = "1.12.0", features = ["attributes"] }
+glob = "0.3.1"
+once_cell = "1.18.0"
+regex = "1.8.4"
+sysconf = "0.3.4"
+serde = { version = "1.0.180", features = ["serde_derive"] }
